+++ onlyoffice-nuxeo-core/src/main/resources/OSGI-INF/callback-service.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0"?>
+<component name="org.onlyoffice.api.CallbackService" version="1.0">
+
+    <implementation class="org.onlyoffice.service.CallbackServiceImpl" />
+
+    <service>
+        <provide interface="org.onlyoffice.api.CallbackService" />
+    </service>
+</component>
