+++ onlyoffice-nuxeo-rest/src/main/java/org/nuxeo/ecm/restapi/server/jaxrs/OnlyofficeObject.java
@@ -20,7 +20,6 @@ package org.nuxeo.ecm.restapi.server.jaxrs;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.URL;
 import java.nio.charset.Charset;
 import java.util.*;
 
@@ -31,31 +30,25 @@ import javax.ws.rs.core.Response.Status;
 
 import org.apache.commons.io.IOUtils;
 import org.json.JSONObject;
-import org.nuxeo.ecm.automation.core.util.DocumentHelper;
 import org.nuxeo.ecm.core.api.*;
-import org.nuxeo.ecm.core.api.blobholder.BlobHolder;
-import org.nuxeo.ecm.core.api.versioning.VersioningService;
-import org.nuxeo.ecm.core.schema.FacetNames;
 import org.nuxeo.ecm.webengine.model.WebObject;
 import org.nuxeo.ecm.webengine.model.exceptions.WebSecurityException;
 import org.nuxeo.ecm.webengine.model.impl.DefaultObject;
 import org.nuxeo.runtime.api.Framework;
+import org.onlyoffice.api.CallbackService;
 import org.onlyoffice.api.SettingsService;
-import org.onlyoffice.utils.ConfigManager;
 import org.onlyoffice.utils.JwtManager;
-import org.onlyoffice.utils.Utils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 @WebObject(type = "onlyoffice")
-public class Callback extends DefaultObject {
+public class OnlyofficeObject extends DefaultObject {
 
-    private static final Logger logger = LoggerFactory.getLogger(Callback.class);
+    private static final Logger logger = LoggerFactory.getLogger(OnlyofficeObject.class);
 
     private JwtManager jwtManager;
     private SettingsService settingsService;
-    private ConfigManager configManager;
-    private Utils utils;
+    private CallbackService callbackService;
 
     @Override
     protected void initialize(Object... args) {
@@ -63,8 +56,7 @@ public class Callback extends DefaultObject {
 
         jwtManager = Framework.getService(JwtManager.class);
         settingsService = Framework.getService(SettingsService.class);
-        configManager = Framework.getService(ConfigManager.class);
-        utils = Framework.getService(Utils.class);
+        callbackService = Framework.getService(CallbackService.class);
     }
 
     @GET
@@ -139,7 +131,7 @@ public class Callback extends DefaultObject {
             CoreSession session = getContext().getCoreSession();
             DocumentModel model = session.getDocument(new IdRef(id));
 
-            processCallback(session, model, json);
+            callbackService.processCallback(session, model, json);
 
         } catch (SecurityException ex) {
             code = Status.UNAUTHORIZED;
@@ -165,68 +157,6 @@ public class Callback extends DefaultObject {
             logger.error("Error while processing callback for " + id, e);
             return Response.status(code).build();
         }
-    }
-
-    private void processCallback(CoreSession session, DocumentModel model, JSONObject json) throws Exception {
-        switch (json.getInt("status")) {
-        case 0:
-            logger.error("ONLYOFFICE has reported that no doc with the specified key can be found");
-            model.removeLock();
-            break;
-        case 1:
-            if (!model.isLocked()) {
-                logger.info("Document open for editing, locking document");
-                model.setLock();
-            } else {
-                logger.debug("Document already locked, another user has entered/exited");
-            }
-            break;
-        case 2:
-            logger.info("Document Updated, changing content");
-            model.removeLock();
-
-            String documentUrl = configManager.replaceDocEditorURLToInnner(json.getString("url"));
-
-            updateDocument(session, model, json.getString("key"), documentUrl);
-            break;
-        case 3:
-            logger.error("ONLYOFFICE has reported that saving the document has failed");
-            model.removeLock();
-            break;
-        case 4:
-            logger.info("No document updates, unlocking node");
-            model.removeLock();
-            break;
-        }
-    }
-
-    private void updateDocument(CoreSession session, DocumentModel model, String changeToken, String url) throws Exception {
-        Blob original = getBlob(model, "file:content");
-        Blob saved = Blobs.createBlob(new URL(url).openStream(), original.getMimeType(), original.getEncoding());
-        saved.setFilename(original.getFilename());
-
-        DocumentHelper.addBlob(model.getProperty("file:content"), saved);
-
-        if (model.hasFacet(FacetNames.VERSIONABLE)) {
-            VersioningOption vo = VersioningOption.MINOR;
-            model.putContextData(VersioningService.VERSIONING_OPTION, vo);
-        }
-
-        model.putContextData(CoreSession.CHANGE_TOKEN, utils.getChangeToken(changeToken));
-
-        session.saveDocument(model);
-        session.save();
-    }
-
-    private Blob getBlob(DocumentModel model, String xpath) {
-        Blob blob = (Blob) model.getPropertyValue(xpath);
-        if (blob == null) {
-            BlobHolder bh = model.getAdapter(BlobHolder.class);
-            if (bh != null) {
-                blob = bh.getBlob();
-            }
-        }
-        return blob;
     }
 
     private void checkAdministrator() {
